requirements.txt
Flask==2.3.2
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
pandas
requests
sidrapy
geobr
geopandas
streamlit
plotly

